requirements.txt
Django==5.0
djangorestframework==3.14.0
python-dotenv==1.0.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9